# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevLog+",
    "description": "Single-user developer journal for technical learning and skill maintenance.\n\nDevLog+ has two engines:\n\n- **Learning engine** — builds a Knowledge Profile from journal entries, quizzes, and feedback. Understands what you know, where you're weak, and what to explore next.\n- **Practice engine** — generates weekly Go micro-projects calibrated to your practical level.\n\n## Core workflow\n\n1. **Journal** — write technical entries (typed or dictated) about what you learn.\n2. **Profile** — the system builds and maintains a Knowledge Profile nightly.\n3. **Quiz** — weekly free-text quizzes probe your understanding.\n4. **Readings** — curated reading recommendations from your allowlisted domains.\n5. **Projects** — weekly Go micro-projects with bugs, features, and refactors.\n6. **Feedback** — thumbs-up/down reactions and feedforward notes steer future content.\n7. **Triage** — items the system can't resolve are surfaced for your clarification.\n",
    "license": {
      "name": "Private — single-user local application"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/journal/entries": {
      "post": {
        "tags": [
          "journal"
        ],
        "summary": "Create journal entry",
        "description": "Create a new technical journal entry.\n\nThe entry content can be typed text or dictated via the browser Web Speech\nAPI.  An initial version is created automatically.  The entry will be\nprocessed by the nightly profile-update pipeline.",
        "operationId": "create_entry_api_v1_journal_entries_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created journal entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error — content is required and must be non-empty"
          }
        }
      },
      "get": {
        "tags": [
          "journal"
        ],
        "summary": "List journal entries",
        "description": "Return journal entries ordered by creation date (most recent first).\n\nSupports cursor-style pagination via `offset` and `limit`.\nEach entry includes its latest content snapshot.",
        "operationId": "list_entries_api_v1_journal_entries_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of entries to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of entries to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum entries to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum entries to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of journal entries, most recent first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalEntryResponse"
                  },
                  "title": "Response List Entries Api V1 Journal Entries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journal/entries/{entry_id}": {
      "get": {
        "tags": [
          "journal"
        ],
        "summary": "Get journal entry with version history",
        "description": "Retrieve a single journal entry including its full version history.\n\nVersions are append-only; the most recent version is the source of truth.",
        "operationId": "get_entry_api_v1_journal_entries__entry_id__get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The journal entry with all historical versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Journal entry not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "journal"
        ],
        "summary": "Edit journal entry",
        "description": "Edit a journal entry.\n\nCreates a new version instead of overwriting — the previous content is\npreserved in the version history.  The new version becomes the source of\ntruth for future profile updates.",
        "operationId": "update_entry_api_v1_journal_entries__entry_id__put",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalEntryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated journal entry (new version created)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Journal entry not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "journal"
        ],
        "summary": "Delete journal entry",
        "description": "Permanently delete a journal entry and all of its versions.",
        "operationId": "delete_entry_api_v1_journal_entries__entry_id__delete",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmation message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Journal entry not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/feedback": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit feedback or feedforward",
        "description": "Submit feedback and/or feedforward on a generated item.\n\nTwo distinct signal types are supported:\n\n- **Feedback** (`reaction`) — a thumbs-up or thumbs-down that corrects\n  the system's current understanding (e.g. “too easy”, “not relevant”).\n- **Feedforward** (`note`) — free-text that shapes what the system should\n  generate next (e.g. “more backend-oriented content”).\n\nAt least one of `reaction` or `note` should be provided.",
        "operationId": "create_feedback_api_v1_feedback_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The recorded feedback entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error — target_type and target_id are required"
          }
        }
      },
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "List feedback",
        "description": "Return all feedback entries ordered by creation date (most recent first).",
        "operationId": "list_feedback_api_v1_feedback_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of entries to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of entries to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum entries to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum entries to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of all feedback entries, most recent first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackResponse"
                  },
                  "title": "Response List Feedback Api V1 Feedback Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeedbackCreate": {
        "properties": {
          "target_type": {
            "$ref": "#/components/schemas/FeedbackTargetType",
            "description": "Type of item receiving feedback: quiz_question, reading, or project"
          },
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id",
            "description": "ID of the target item (question, reading, or project)"
          },
          "reaction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackReaction"
              },
              {
                "type": "null"
              }
            ],
            "description": "Quick reaction: thumbs_up or thumbs_down (feedback signal)"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Free-text feedforward note that shapes future generation (e.g. 'more backend-oriented content', 'harder debugging tasks')",
            "examples": [
              "This was too easy — I'd prefer more advanced concurrency challenges."
            ]
          }
        },
        "type": "object",
        "required": [
          "target_type",
          "target_id"
        ],
        "title": "FeedbackCreate",
        "description": "Submit feedback and/or feedforward on a generated item.\n\nAt least one of `reaction` or `note` should be provided."
      },
      "FeedbackReaction": {
        "type": "string",
        "enum": [
          "thumbs_up",
          "thumbs_down"
        ],
        "title": "FeedbackReaction"
      },
      "FeedbackResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Feedback entry ID"
          },
          "target_type": {
            "$ref": "#/components/schemas/FeedbackTargetType",
            "description": "Type of item this feedback targets"
          },
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id",
            "description": "ID of the targeted item"
          },
          "reaction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackReaction"
              },
              {
                "type": "null"
              }
            ],
            "description": "Thumbs-up / thumbs-down reaction, if provided"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Free-text feedforward note, if provided"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the feedback was submitted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "target_type",
          "target_id",
          "reaction",
          "note",
          "created_at"
        ],
        "title": "FeedbackResponse",
        "description": "A recorded feedback/feedforward entry."
      },
      "FeedbackTargetType": {
        "type": "string",
        "enum": [
          "quiz_question",
          "reading",
          "project",
          "project_task"
        ],
        "title": "FeedbackTargetType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JournalEntryCreate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Optional short title for the entry",
            "examples": [
              "Learning Go concurrency patterns"
            ]
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "Entry content (text only, no HTML)",
            "examples": [
              "Today I explored Go channels and goroutines. The select statement is particularly powerful for multiplexing channel operations."
            ]
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "JournalEntryCreate",
        "description": "Create a new journal entry.\n\nContent is text-only and can be typed or dictated via the browser\nWeb Speech API.  Title is optional."
      },
      "JournalEntryDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Entry ID"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Optional title"
          },
          "is_processed": {
            "type": "boolean",
            "title": "Is Processed",
            "description": "Whether the nightly pipeline has processed this entry"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At",
            "description": "When the entry was last processed by the pipeline"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the entry was first created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "When the entry was last modified"
          },
          "current_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Content",
            "description": "Content from the most recent version"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/JournalEntryVersionResponse"
            },
            "type": "array",
            "title": "Versions",
            "description": "All versions of this entry, ordered by version number",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "is_processed",
          "processed_at",
          "created_at",
          "updated_at"
        ],
        "title": "JournalEntryDetailResponse",
        "description": "Full journal entry with append-only version history."
      },
      "JournalEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Entry ID"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Optional title"
          },
          "is_processed": {
            "type": "boolean",
            "title": "Is Processed",
            "description": "Whether the nightly pipeline has processed this entry"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At",
            "description": "When the entry was last processed by the pipeline"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the entry was first created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "When the entry was last modified"
          },
          "current_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Content",
            "description": "Content from the most recent version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "is_processed",
          "processed_at",
          "created_at",
          "updated_at"
        ],
        "title": "JournalEntryResponse",
        "description": "Journal entry summary (without version history)."
      },
      "JournalEntryUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Updated title (null leaves unchanged)"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "Updated content (will become the new current version)"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "JournalEntryUpdate",
        "description": "Edit an existing entry — creates a new version, preserving history."
      },
      "JournalEntryVersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Version ID"
          },
          "entry_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entry Id",
            "description": "Parent journal entry ID"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Full text content of this version"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number",
            "description": "Sequential version number (1-based)"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current",
            "description": "Whether this is the latest version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When this version was created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entry_id",
          "content",
          "version_number",
          "is_current",
          "created_at"
        ],
        "title": "JournalEntryVersionResponse",
        "description": "A single version of a journal entry's content."
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable result message",
            "examples": [
              "Entry deleted"
            ]
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Generic message response for operations that return no entity."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}